+++ src/mpd/ThreadedMinimumPairwiseDistance.java
@@ -3,7 +3,7 @@ package mpd;
 public class ThreadedMinimumPairwiseDistance implements MinimumPairwiseDistance {
 
     @Override
-    public long minimumPairwiseDistance(int[] values) {
+    public long minimumPairwiseDistance(long[] values) {
         throw new UnsupportedOperationException();
     }
